requirements.txt
Flask==0.12.2
gunicorn==19.6.0
line-bot-sdk==1.3.0

numpy == 1.16.3
pandas == 0.24.1
Pillow == 6.0.0
requests == 2.23.0
urllib3 == 1.24.3
tensorflow-cpu==2.3.1
